IEntity are now using glm::vec2 to stock positions

+++ include/IEntity.hpp
@@ -23,10 +23,10 @@ public:
     };
 
   virtual ~IEntity() {}
-  virtual int getPosX() const = 0;
-  virtual int getPosY() const = 0;
-  virtual void setPosX(const int) = 0;
-  virtual void setPosY(const int) = 0;
+  virtual float getPosX() const = 0;
+  virtual float getPosY() const = 0;
+  virtual void setPosX(const float) = 0;
+  virtual void setPosY(const float) = 0;
   virtual AObject *getObj() = 0;
 };
